+++ tests/all_tests.py
@@ -18,7 +18,7 @@ __copyright__ = "Copyright 2011-2013, The PICRUSt Project"
 __credits__ = ["Morgan Langille","Jose Carlos Clemente Litran", "Daniel McDonald",
                "Greg Caporaso", "Jai Ram Rideout"]
 __license__ = "GPL"
-__version__ = "1.1.3"
+__version__ = "1.1.4"
 __maintainer__ = "Greg Caporaso"
 __email__ = "gregcaporaso@gmail.com"